fix looping video bug when user changes time on ended video
Fixed a weird bug where the client ends up in an infinite loop. This was caused by the user trying to change the time when the video had ended. This caused the player to try to play the video, the server tries to end it, the player tries to play it again, etc.