add session ID filter in RoomController.Join
Added a whitelist of session IDs allowed in a given room. When a user navigates to a room page, the given room adds the session ID to its whitelist, which means that person can send join requests and have it be successful.

This also means that anyone who tries to send a join request without being in the room will have it fail.