fix buggy logic in RoomManager destroying rooms
Beforehand, the room manager would always check the room age before checking if the room was empty. Then, if the room was empty, it would be destroyed.

Now, the room manager first checks if the room is empty, and then afterwards checks if the room is too young to destroy. If the room is old enough, it is then destroyed.