fully implement client side join request, partially add server updates
Made it so the client now successfully handles join requests, it sends a join request, and then receives and interprets a response from the server. It then sets the state of the video and populates the user list.

The client now has a method for handling updates from the server - i.e. things the server sends without a request being sent by the client first. This is not fully implemented yet and has no functionality. It just successfully checks the type of update being sent.