fix socket closed exception in ConnectionManager
An exception would happen when a user left the room and the server tried to send an update to that user. The exception was basically saying that we were trying to send data through a socket that was closed. This was fixed by adding a try catch block to ConnectionManager.SendUpdateToAll and SendUpdateToAllExcept. Now users who have closed sockets will be disconnected from the room.