fix bug where video time was no longer regularly being updated in the room
There was a bug where the video time was no longer being regularly updated in the room. This is because of the conversion to WebSockets. The video time used to get updated all the time because the users were constantly sending update requests to the room, but now with WebSockets, users are no longer constantly sending updates.

This bug was fixed by reintroducing periodic tasks in the room that execute regularly, and adding a periodic task that updates the video time.

The periodic task to kick out inactive users has been commented out.