add "add to playlist" functionality
Users can now add to the playlist. But there is a bug where the playlist will show "video title not found" (etc) because the update is so fast that the oembed API doesn't have time to set the video title properly. SendUpdateToAll should be called when noembed successfully finds the video data and edits the playlist.