add play playlist video functionality + bug fix
added functionality for the user to click on a playlist item and the room playing the relevant video.

Fixed the same bug in the last commit where the final curly bracket in the json would not be present after being sent to the client. The fix also involved removing the creation of an ArraySegment, this time in ConnectionManager.cs