fix client side reconnect behaviour
Fixed issues with the way the client behaves when attempting to reconnect. The logic was wrong in that the method tryReconnect() was calling upon itself every 5 seconds with setTimeout(...), not leaving enough time for a real connection attempt to be made.So the client kept trying to reconnect but those reconnection attempts would get "doubled up" by the method being called again and again. Now, the way it works is that tryReconnect() is always called upon socket.onclose event, because socket.onclose gets called whenever the connection attempt fails. This is still limited to 50 connection attempts, just the tracking of the number of connection attempts is now done in the socket.onclose instead of inside tryReconnect()