fix behaviour where rooms could be destroyed too quickly
Fixed a behaviour where rooms could be destroyed too quickly after everyone has left. Occasionally, even if you just refresh the page, the room could be destroyed if you were the only one in the room. This has been fixed by adding a timer inside the room that counts to 2 minutes, at which point the room will be tagged as destroyable. The RoomManager will destroy any room that has been tagged as destroyable.

This means that rooms can be empty for 2 minutes before they are destroyed.