try make loading screen appear before page load
Attempted to fix an issue where the loading screen would sometimes appear a little late, with the browser showing the loading of the page before the loading screen itself. I added a <script> tag to the html which makes the loading screen appear in order to try and make the loading screen appear as soon as the div for it is loaded.